network: Move ethtool to pkg/network/link
